rename background to canvasBackground
