--- common/recoil/background/background.atom.ts
@@ -1,9 +1,0 @@
-import { atom } from "recoil";
-
-export const backgroundAtom = atom<{ mode: "dark" | "light"; lines: boolean }>({
-  key: "bg",
-  default: {
-    mode: "light",
-    lines: true,
-  },
-});
