--- common/recoil/background/index.ts
@@ -1,6 +1,0 @@
-import { backgroundAtom } from "./background.atom";
-import { useBackground, useSetBackground } from "./background.hook";
-
-export default backgroundAtom;
-
-export { useBackground, useSetBackground };
