+++ common/recoil/canvasBackground/index.ts
@@ -1,0 +1,6 @@
+import { canvasBackgroundAtom } from "./canvasBackground.atom";
+import { useBackground, useSetBackground } from "./canvasBackground.hook";
+
+export default canvasBackgroundAtom;
+
+export { useBackground, useSetBackground };
